+++ src/utils/govcyValidator.mjs
@@ -263,8 +263,9 @@ export function validateFormElements(elements, formData, pageUrl) {
         formData[`${field.params.name}_day`]]
           .filter(Boolean) // Remove empty values
           .join("-") // Join remaining parts
-        : (field.element === "fileInput") // Handle fileInput
-          ? formData[`${field.params.name}Attachment`] || ""
+        // unneeded handle of `Attachment` at the end
+        // : (field.element === "fileInput") // Handle fileInput
+        //   ? formData[`${field.params.name}Attachment`] || ""
         : formData[field.params.name] || ""; // Get submitted value
 
       //Autocheck: check for "checkboxes", "radios", "select" if `fieldValue` is one of the `field.params.items
@@ -313,7 +314,9 @@ export function validateFormElements(elements, formData, pageUrl) {
                     .filter(Boolean) // Remove empty values
                     .join("-") // Join remaining parts
                   : (conditionalElement.element === "fileInput") // Handle fileInput
-                      ? formData[`${conditionalElement.params.name}Attachment`] || ""
+                      // unneeded handle of `Attachment` at the end
+                      // ? formData[`${conditionalElement.params.name}Attachment`] || ""
+                      ? formData[`${conditionalElement.params.name}`] || ""
                   : formData[conditionalElement.params.name] || ""; // Get submitted value
 
                 //Autocheck: check for "checkboxes", "radios", "select" if `fieldValue` is one of the `field.params.items`